Improvments on Postgrecode, Refactorings

+++ OpenPrismNode.Sync/PostgresModels/Metadata.cs
@@ -5,9 +5,9 @@
 /// </summary>
 public class Metadata
 {
-    public long id { get; set; }
+    public int id { get; set; }
     public decimal key { get; set; }
     public string json { get; set; } = String.Empty;
     public byte[] bytes { get; set; } = Array.Empty<byte>();
-    public long tx_id { get; set; }
+    public int tx_id { get; set; }
 }
